Move chunk "physics" out of engine
Depends on a bunch of game state
